some refactoring, dependency injection

+++ backend/src/main/java/com/rewardapp/backend/controllers/RecyclingCenterController.java
@@ -1,13 +1,11 @@
 package com.rewardapp.backend.controllers;
 
-import com.rewardapp.backend.entities.Session;
-import com.rewardapp.backend.models.RecyclingCenterModel;
+import com.rewardapp.backend.models.RecyclingCenter;
+import com.rewardapp.backend.services.AuthService;
 import com.rewardapp.backend.services.RecyclingCenterService;
-import com.rewardapp.backend.services.SessionService;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.hateoas.CollectionModel;
-import org.springframework.hateoas.EntityModel;
 import org.springframework.hateoas.RepresentationModel;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpStatus;
@@ -21,46 +19,49 @@ import java.util.List;
 @RequiredArgsConstructor
 public class RecyclingCenterController {
     private final RecyclingCenterService recyclingCenterService;
-    private final SessionService sessionService;
+    private final AuthService authService;
 
     @GetMapping("/{id}")
-    public ResponseEntity<RepresentationModel> get(@PathVariable(name = "id") Long id, HttpServletRequest request) {
-        Session session = sessionService.validateRequest(request);
-        RecyclingCenterModel model = recyclingCenterService.findRecyclingCenterById(id);
+    public ResponseEntity<RepresentationModel<RecyclingCenter>> get(@PathVariable(name = "id") Long id, HttpServletRequest request) {
+        authService.validateRequest(request);
 
         return ResponseEntity
                 .status(HttpStatus.OK)
-                .body(EntityModel.of(model));
+                .body(recyclingCenterService.findRecyclingCenterById(id));
     }
 
     @GetMapping("/all")
-    public ResponseEntity<CollectionModel> getAll(HttpServletRequest request) {
-        Session session = sessionService.validateRequest(request);
+    public ResponseEntity<CollectionModel<RecyclingCenter>> getAll(HttpServletRequest request) {
+        authService.validateRequest(request);
 
-        List<RecyclingCenterModel> recyclingCenters = recyclingCenterService.getAll();
-
-        CollectionModel<EntityModel<RecyclingCenterModel>> collectionModel = CollectionModel.wrap(recyclingCenters);
+        List<RecyclingCenter> recyclingCenters = recyclingCenterService.getAll();
 
         return ResponseEntity
                 .status(HttpStatus.OK)
-                .body(collectionModel);
+                .body(CollectionModel.of(recyclingCenters));
     }
 
     @PostMapping
-    public ResponseEntity<RepresentationModel> create(@RequestBody RecyclingCenterModel recyclingCenterModel, HttpServletRequest request) {
-        Session session = sessionService.validateAdminRequest(request);
-
-        recyclingCenterModel = recyclingCenterService.create(recyclingCenterModel);
+    public ResponseEntity<RepresentationModel<RecyclingCenter>> create(@RequestBody RecyclingCenter recyclingCenter, HttpServletRequest request) {
+        authService.validateAdminRequest(request);
 
         return ResponseEntity
                 .status(HttpStatus.OK)
-                .body(recyclingCenterModel);
+                .body(recyclingCenterService.create(recyclingCenter));
     }
 
+    @PatchMapping("/{id}")
+    public ResponseEntity<RepresentationModel<RecyclingCenter>> update(@PathVariable("id") Long id, @RequestBody RecyclingCenter recyclingCenter, HttpServletRequest request) {
+        authService.validateAdminRequest(request);
+
+        return ResponseEntity
+                .status(HttpStatus.OK)
+                .body(recyclingCenterService.update(id, recyclingCenter));
+    }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<HttpEntity<Boolean>> delete(@PathVariable("id") Long id, HttpServletRequest request) {
-        Session session = sessionService.validateAdminRequest(request);
+        authService.validateAdminRequest(request);
 
         Boolean status = recyclingCenterService.deleteById(id);
         HttpEntity<Boolean> httpEntity = new HttpEntity<>(status);
@@ -68,16 +69,5 @@ public class RecyclingCenterController {
         return ResponseEntity
                 .status(status ? HttpStatus.OK : HttpStatus.CONFLICT)
                 .body(httpEntity);
-    }
-
-    @PatchMapping("/{id}")
-    public ResponseEntity<RepresentationModel> update(@PathVariable("id") Long id, @RequestBody RecyclingCenterModel recyclingCenterModel, HttpServletRequest request) {
-        Session session = sessionService.validateAdminRequest(request);
-
-        recyclingCenterModel = recyclingCenterService.update(id, recyclingCenterModel);
-
-        return ResponseEntity
-                .status(HttpStatus.OK)
-                .body(EntityModel.of(recyclingCenterModel));
     }
 }
